+++ src/schemathesis/_hypothesis.py
@@ -1,11 +1,14 @@
 """Provide strategies for given endpoint(s) definition."""
 import asyncio
+import re
 from functools import partial
-from typing import Any, Callable, Generator, Optional
+from typing import Any, Callable, Dict, Generator, Optional
 
 import hypothesis
 import hypothesis.strategies as st
 from hypothesis_jsonschema import from_schema
+from requests.exceptions import InvalidHeader  # type: ignore
+from requests.utils import check_header_validity  # type: ignore
 
 from ._compat import handle_warnings
 from .models import Case, Endpoint
@@ -62,6 +65,42 @@ def add_examples(test: Callable, endpoint: Endpoint) -> Callable:
     return test
 
 
+# Adapted from http.client._is_illegal_header_value
+INVALID_HEADER_RE = re.compile(r"\n(?![ \t])|\r(?![ \t\n])")
+
+
+def _is_latin_1_encodable(value: str) -> bool:
+    """Header values are encoded to latin-1 before sending.
+
+    We need to generate valid payload.
+    """
+    try:
+        value.encode("latin-1")
+        return True
+    except UnicodeEncodeError:
+        return False
+
+
+def _has_invalid_characters(name: str, value: str) -> bool:
+    try:
+        check_header_validity((name, value))
+        return bool(INVALID_HEADER_RE.search(value))
+    except InvalidHeader:
+        return True
+
+
+def is_valid_header(headers: Optional[Dict[str, str]]) -> bool:
+    """Verify if the generated headers are valid."""
+    if headers is None:
+        return True
+    for name, value in headers.items():
+        if not _is_latin_1_encodable(value):
+            return False
+        if _has_invalid_characters(name, value):
+            return False
+    return True
+
+
 def get_case_strategy(endpoint: Endpoint) -> st.SearchStrategy:
     """Create a strategy for a complete test case.
 
@@ -70,7 +109,7 @@ def get_case_strategy(endpoint: Endpoint) -> st.SearchStrategy:
     return st.builds(
         partial(Case, path=endpoint.path, method=endpoint.method, base_url=endpoint.base_url),
         path_parameters=from_schema(endpoint.path_parameters),
-        headers=from_schema(endpoint.headers),
+        headers=from_schema(endpoint.headers).filter(is_valid_header),  # type: ignore
         cookies=from_schema(endpoint.cookies),
         query=from_schema(endpoint.query),
         body=from_schema(endpoint.body),